Refactor MessageDetail and Timeline components for improved structure and functionality
- Removed unused imports and code from MessageDetailPage.vue
- Simplified message detail rendering and error handling
- Enhanced user interface with better loading and error states
- Consolidated message posting logic in TimelinePage.vue using MessageForm and MessageList components
- Improved image upload handling and error management in TimelinePage.vue
- Updated styles for better responsiveness and visual consistency
