+++ frontend/src/pages/MessageDetail/MessageDetailPage.vue
@@ -1,11 +1,6 @@
 <script lang="ts" setup>
-import {
-  useAddReaction,
-  useCreateMessage,
-  useFormState,
-  useMessageDetail,
-  useRemoveReaction,
-} from '@/lib/composables'
+import UserIcon from '@/components/UserIcon.vue'
+import { useMessageDetail } from '@/lib/composables'
 import { computed } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 
@@ -18,55 +13,21 @@ const messageId = computed(() => route.params.id as string)
 // メッセージ詳細取得
 const { data: message, isLoading, error } = useMessageDetail(messageId)
 
-// リプライ作成
-const createMessageMutation = useCreateMessage()
-
-// リアクション機能
-const addReactionMutation = useAddReaction()
-const removeReactionMutation = useRemoveReaction()
-
-// リプライフォーム状態管理
-const {
-  formData: replyForm,
-  resetForm,
-  setSubmitting,
-  isSubmitting,
-} = useFormState({
-  content: '',
-})
-
-// リアクション切り替え処理
-const toggleReaction = async () => {
-  if (!messageId.value || !message.value) return
-
-  try {
-    if (message.value.reactions.myReaction) {
-      await removeReactionMutation.mutateAsync(messageId.value)
-    } else {
-      await addReactionMutation.mutateAsync(messageId.value)
-    }
-  } catch (error) {
-    console.error('Failed to toggle reaction:', error)
-  }
+// フォーマット済み作成日時
+const formatDate = (dateString: string) => {
+  const date = new Date(dateString)
+  return date.toLocaleDateString('ja-JP', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit',
+  })
 }
 
-// リプライ投稿処理
-const handleReplySubmit = async () => {
-  if (!replyForm.value.content.trim() || !messageId.value) return
-
-  setSubmitting(true)
-
-  try {
-    await createMessageMutation.mutateAsync({
-      content: replyForm.value.content,
-      repliesTo: messageId.value,
-    })
-    resetForm()
-  } catch (error) {
-    console.error('Failed to create reply:', error)
-  } finally {
-    setSubmitting(false)
-  }
+// ユーザー詳細ページへの遷移
+const goToUserDetail = (traqId: string) => {
+  router.push(`/users/${traqId}`)
 }
 
 // 戻る
@@ -76,179 +37,238 @@ const goBack = () => {
 </script>
 
 <template>
-  <main :class="$style.messageDetail">
+  <div :class="$style.messageDetail">
     <div :class="$style.container">
       <!-- ヘッダー -->
-      <div :class="$style.header">
+      <header :class="$style.header">
         <button :class="$style.backButton" @click="goBack">← 戻る</button>
-        <h1>メッセージの詳細</h1>
+        <h1 :class="$style.title">メッセージ詳細</h1>
+      </header>
+
+      <!-- ローディング状態 -->
+      <div v-if="isLoading" :class="$style.loading">
+        <div :class="$style.spinner"></div>
+        <p>メッセージを読み込み中...</p>
       </div>
 
-      <!-- エラー表示 -->
-      <div v-if="error" :class="$style.error">
-        <p>メッセージの取得に失敗しました</p>
-        <button @click="goBack">戻る</button>
+      <!-- エラー状態 -->
+      <div v-else-if="error" :class="$style.error">
+        <div :class="$style.errorIcon">⚠️</div>
+        <h2>メッセージの取得に失敗しました</h2>
+        <p>しばらく待ってから再度お試しください。</p>
       </div>
 
-      <!-- ローディング表示 -->
-      <div v-else-if="isLoading" :class="$style.loading">
-        <p>読み込み中...</p>
+      <!-- メッセージが見つからない -->
+      <div v-else-if="!message" :class="$style.notFound">
+        <div :class="$style.errorIcon">📭</div>
+        <h2>メッセージが見つかりません</h2>
+        <p>指定されたメッセージは削除されたか、存在しません。</p>
       </div>
 
-      <!-- メッセージ詳細 -->
-      <div v-else-if="message" :class="$style.content">
+      <!-- メッセージ内容 -->
+      <main v-else :class="$style.content">
         <!-- メインメッセージ -->
-        <div :class="$style.mainMessage">
+        <article :class="$style.mainMessage">
           <div :class="$style.messageHeader">
-            <div :class="$style.userInfo">
-              <strong>{{ message.author || 'Unknown User' }}</strong>
-              <span :class="$style.timestamp">{{
-                new Date(message.createdAt).toLocaleString()
-              }}</span>
+            <UserIcon
+              :traq-id="message.author"
+              size="lg"
+              clickable
+              @click="goToUserDetail(message.author)"
+            />
+            <div :class="$style.messageInfo">
+              <button :class="$style.authorName" @click="goToUserDetail(message.author)">
+                @{{ message.author }}
+              </button>
+              <time :class="$style.timestamp" :datetime="message.createdAt">
+                {{ formatDate(message.createdAt) }}
+              </time>
             </div>
           </div>
+
           <div :class="$style.messageContent">
-            {{ message.content }}
+            <p :class="$style.messageText">
+              {{ message.content }}
+            </p>
+
+            <!-- 画像表示 -->
+            <div v-if="message.imageId" :class="$style.imageContainer">
+              <img
+                :src="`/api/images/${message.imageId}`"
+                :alt="'添付画像'"
+                :class="$style.messageImage"
+                loading="lazy"
+              />
+            </div>
           </div>
 
           <!-- リアクション表示（読み取り専用） -->
-          <div v-if="message.reactions && message.reactions.count > 0" :class="$style.reactions">
-            <div :class="$style.reaction">
-              👍 {{ message.reactions.count }}
-              <span v-if="message.reactions.myReaction" :class="$style.myReaction"
-                >（あなたがリアクション済み）</span
-              >
-            </div>
-          </div>
-
-          <!-- リアクションボタン -->
-          <div :class="$style.reactionActions">
-            <button
-              :class="[$style.reactionButton, { [$style.active]: message.reactions.myReaction }]"
-              @click="toggleReaction"
-              :disabled="
-                addReactionMutation.isPending.value || removeReactionMutation.isPending.value
-              "
+          <div v-if="message.reactions.count > 0" :class="$style.reactions">
+            <div
+              :class="[$style.reactionItem, { [$style.myReaction]: message.reactions.myReaction }]"
             >
-              👍 {{ message.reactions.myReaction ? 'リアクション済み' : 'リアクション' }}
-            </button>
-          </div>
-        </div>
-
-        <!-- リプライフォーム -->
-        <div :class="$style.replyForm">
-          <h3>このメッセージにリプライ</h3>
-          <form @submit.prevent="handleReplySubmit">
-            <div :class="$style.formGroup">
-              <textarea
-                v-model="replyForm.content"
-                :class="$style.textarea"
-                placeholder="リプライ内容を入力..."
-                rows="3"
-                :disabled="isSubmitting"
-              />
-            </div>
-            <div :class="$style.formActions">
-              <button
-                type="submit"
-                :class="$style.submitButton"
-                :disabled="isSubmitting || !replyForm.content.trim()"
-              >
-                {{ isSubmitting ? 'リプライ中...' : 'リプライ' }}
-              </button>
-            </div>
-          </form>
-        </div>
-
-        <!-- リプライ一覧（読み取り専用） -->
-        <div v-if="message?.replies && message.replies.length > 0" :class="$style.replies">
-          <h3>リプライ ({{ message.replies.length }})</h3>
-          <div v-for="reply in message.replies" :key="reply.id" :class="$style.reply">
-            <div :class="$style.replyHeader">
-              <strong>{{ reply.author || 'Unknown User' }}</strong>
-              <span :class="$style.timestamp">{{
-                new Date(reply.createdAt).toLocaleString()
-              }}</span>
-            </div>
-            <div :class="$style.replyContent">
-              {{ reply.content }}
+              <span :class="$style.emoji">👍</span>
+              <span :class="$style.count">{{ message.reactions.count }}</span>
             </div>
           </div>
-        </div>
-      </div>
+        </article>
 
-      <!-- メッセージが見つからない場合 -->
-      <div v-else :class="$style.notFound">
-        <p>メッセージが見つかりません</p>
-        <button @click="goBack">戻る</button>
-      </div>
+        <!-- 返信一覧 -->
+        <section v-if="message.replies && message.replies.length > 0" :class="$style.replies">
+          <h2 :class="$style.repliesTitle">返信 ({{ message.replies.length }})</h2>
+          <div :class="$style.repliesList">
+            <article v-for="reply in message.replies" :key="reply.id" :class="$style.reply">
+              <div :class="$style.messageHeader">
+                <UserIcon
+                  :traq-id="reply.author"
+                  size="md"
+                  clickable
+                  @click="goToUserDetail(reply.author)"
+                />
+                <div :class="$style.messageInfo">
+                  <button :class="$style.authorName" @click="goToUserDetail(reply.author)">
+                    @{{ reply.author }}
+                  </button>
+                  <time :class="$style.timestamp" :datetime="reply.createdAt">
+                    {{ formatDate(reply.createdAt) }}
+                  </time>
+                </div>
+              </div>
+
+              <div :class="$style.messageContent">
+                <p :class="$style.messageText">
+                  {{ reply.content }}
+                </p>
+
+                <!-- 返信の画像表示 -->
+                <div v-if="reply.images" :class="$style.imageContainer">
+                  <img
+                    :src="`/api/images/${reply.images}`"
+                    :alt="'添付画像'"
+                    :class="$style.messageImage"
+                    loading="lazy"
+                  />
+                </div>
+              </div>
+
+              <!-- 返信のリアクション表示（読み取り専用） -->
+              <div v-if="reply.reactions.count > 0" :class="$style.reactions">
+                <div
+                  :class="[
+                    $style.reactionItem,
+                    { [$style.myReaction]: reply.reactions.myReaction },
+                  ]"
+                >
+                  <span :class="$style.emoji">👍</span>
+                  <span :class="$style.count">{{ reply.reactions.count }}</span>
+                </div>
+              </div>
+            </article>
+          </div>
+        </section>
+      </main>
     </div>
-  </main>
+  </div>
 </template>
 
 <style lang="scss" module>
 .messageDetail {
-  padding: 1rem;
-  max-width: 800px;
-  margin: 0 auto;
+  min-height: 100vh;
+  background-color: var(--color-background);
 }
 
 .container {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 0 1rem;
 }
 
 .header {
   display: flex;
   align-items: center;
   gap: 1rem;
-
-  h1 {
-    margin: 0;
-    color: var(--color-text-primary);
-  }
+  padding: 1rem 0;
+  border-bottom: 1px solid var(--color-border-light);
+  margin-bottom: 2rem;
 }
 
 .backButton {
   background: none;
-  border: 1px solid var(--color-border-medium);
-  border-radius: 4px;
-  padding: 0.5rem 1rem;
-  cursor: pointer;
+  border: none;
   color: var(--color-primary-600);
+  cursor: pointer;
+  font-size: 0.875rem;
+  padding: 0.5rem;
+  border-radius: 0.25rem;
+  transition: background-color 0.2s ease;
 
   &:hover {
-    background: var(--color-surface-variant);
+    background-color: var(--color-primary-50);
   }
+}
+
+.title {
+  margin: 0;
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: var(--color-text-primary);
+}
+
+.loading,
+.error,
+.notFound {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 4rem 2rem;
+  text-align: center;
+  color: var(--color-text-secondary);
+}
+
+.loading {
+  gap: 1rem;
 }
 
 .error,
 .notFound {
-  background: var(--color-error-50);
-  color: var(--color-error-700);
-  padding: 1.5rem;
-  border-radius: 8px;
-  text-align: center;
+  gap: 0.5rem;
+}
 
-  button {
-    background: var(--color-error-600);
-    color: var(--color-text-inverse);
-    border: none;
-    border-radius: 4px;
-    padding: 0.5rem 1rem;
-    margin-top: 1rem;
-    cursor: pointer;
+.spinner {
+  width: 2rem;
+  height: 2rem;
+  border: 2px solid var(--color-border-light);
+  border-top: 2px solid var(--color-primary-500);
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+}
 
-    &:hover {
-      background: var(--color-error-700);
-    }
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
   }
 }
 
-.loading {
-  text-align: center;
-  padding: 3rem;
+.errorIcon {
+  font-size: 3rem;
+  margin-bottom: 1rem;
+}
+
+.error h2,
+.notFound h2 {
+  margin: 0 0 0.5rem 0;
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: var(--color-text-primary);
+}
+
+.error p,
+.notFound p {
+  margin: 0;
   color: var(--color-text-secondary);
 }
 
@@ -258,176 +278,148 @@ const goBack = () => {
   gap: 2rem;
 }
 
-.mainMessage {
-  background: var(--color-surface);
-  border-radius: 8px;
+.mainMessage,
+.reply {
+  background-color: var(--color-surface);
+  border-radius: 0.75rem;
   padding: 1.5rem;
-  box-shadow: 0 2px 4px var(--color-shadow-light);
+  box-shadow: 0 1px 3px var(--color-shadow-light);
 }
 
 .messageHeader {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
   margin-bottom: 1rem;
 }
 
-.userInfo {
+.messageInfo {
   display: flex;
-  align-items: center;
-  gap: 1rem;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.authorName {
+  background: none;
+  border: none;
+  padding: 0;
+  font-size: 1rem;
+  font-weight: 600;
+  color: var(--color-primary-600);
+  cursor: pointer;
+  text-align: left;
+
+  &:hover {
+    color: var(--color-primary-700);
+    text-decoration: underline;
+  }
 }
 
 .timestamp {
-  color: var(--color-text-secondary);
   font-size: 0.875rem;
+  color: var(--color-text-secondary);
 }
 
 .messageContent {
-  font-size: 1.1rem;
-  line-height: 1.6;
   margin-bottom: 1rem;
+}
+
+.messageText {
+  margin: 0;
+  line-height: 1.6;
+  color: var(--color-text-primary);
   white-space: pre-wrap;
+  word-wrap: break-word;
+}
+
+.imageContainer {
+  margin-top: 1rem;
+}
+
+.messageImage {
+  max-width: 100%;
+  max-height: 20rem;
+  border-radius: 0.5rem;
+  border: 1px solid var(--color-border-light);
 }
 
 .reactions {
   display: flex;
   gap: 0.5rem;
-  margin-bottom: 1rem;
+  flex-wrap: wrap;
 }
 
-.reaction {
-  background: var(--color-surface-variant);
-  border: 1px solid var(--color-border-light);
-  border-radius: 20px;
-  padding: 0.25rem 0.75rem;
-  font-size: 0.875rem;
-  cursor: pointer;
+.reactionItem {
   display: flex;
   align-items: center;
-  gap: 0.5rem;
-
-  &:hover {
-    background: var(--color-surface);
-  }
-}
-
-.myReaction {
-  font-size: 0.75rem;
-  color: var(--color-primary-600);
-  font-style: italic;
-}
-
-.reactionActions {
-  display: flex;
   gap: 0.25rem;
-  margin-top: 1rem;
+  padding: 0.25rem 0.5rem;
+  background-color: var(--color-surface-variant);
+  border: 1px solid var(--color-border-light);
+  border-radius: 1rem;
+  font-size: 0.75rem;
+
+  &.myReaction {
+    background-color: var(--color-primary-50);
+    border-color: var(--color-primary-200);
+    color: var(--color-primary-700);
+
+    [data-theme='dark'] & {
+      background-color: var(--color-primary-900);
+      border-color: var(--color-primary-800);
+      color: var(--color-primary-300);
+    }
+  }
 }
 
-.emojiButton {
-  background: none;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  padding: 0.25rem 0.5rem;
-  font-size: 1.2rem;
-  cursor: pointer;
+.emoji {
+  font-size: 1rem;
+  line-height: 1;
+}
 
-  &:hover {
-    background: #f8f9fa;
-  }
+.count {
+  font-weight: 500;
+  min-width: 1rem;
+  text-align: center;
 }
 
 .replies {
-  h3 {
-    margin-bottom: 1rem;
-    color: var(--color-text-primary);
-  }
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.repliesTitle {
+  margin: 0 0 1rem 0;
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: var(--color-text-primary);
+}
+
+.repliesList {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
 }
 
 .reply {
-  background: var(--color-surface-variant);
-  border-radius: 8px;
-  padding: 1rem;
-  margin-bottom: 1rem;
+  margin-left: 1rem;
+  border-left: 3px solid var(--color-primary-200);
 }
 
-.replyHeader {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  margin-bottom: 0.5rem;
-}
-
-.replyContent {
-  line-height: 1.5;
-  white-space: pre-wrap;
-}
-
-.replyForm {
-  background: var(--color-surface);
-  border-radius: 8px;
-  padding: 1.5rem;
-  box-shadow: 0 2px 4px var(--color-shadow-light);
-}
-
-.formGroup {
-  margin-bottom: 1rem;
-}
-
-.textarea {
-  width: 100%;
-  border: 1px solid var(--color-border-medium);
-  border-radius: 4px;
-  padding: 0.75rem;
-  font-size: 1rem;
-  resize: vertical;
-  background-color: var(--color-surface);
-  color: var(--color-text-primary);
-
-  &:focus {
-    outline: none;
-    border-color: var(--color-primary-500);
-    box-shadow: 0 0 0 2px var(--color-primary-200);
+/* レスポンシブ対応 */
+@media (max-width: 768px) {
+  .container {
+    padding: 0 0.5rem;
   }
 
-  &:disabled {
-    background-color: var(--color-surface-variant);
-    cursor: not-allowed;
-  }
-}
-
-.formActions {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.submitButton {
-  background: var(--color-primary-600);
-  color: var(--color-text-inverse);
-  border: none;
-  border-radius: 4px;
-  padding: 0.5rem 1.5rem;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.2s;
-
-  &:hover:not(:disabled) {
-    background: var(--color-primary-700);
+  .mainMessage,
+  .reply {
+    padding: 1rem;
   }
 
-  &:disabled {
-    background: var(--color-gray-500);
-    cursor: not-allowed;
-  }
-}
-
-.reactionButton {
-  background: none;
-  border: 1px solid var(--color-border-medium);
-  border-radius: 4px;
-  padding: 0.25rem 0.5rem;
-  font-size: 1.2rem;
-  cursor: pointer;
-
-  &:hover {
-    background: var(--color-surface-variant);
+  .reply {
+    margin-left: 0.5rem;
   }
 }
 </style>
